Devops: Add error handling if fork requires secret of upstream
The if-condition that checks if `github.repository` is the upstream repo
does not consider a branch on a fork is requested to be merged to the
upstream. In this case the workflow will run but will fail because the
access to the secrets is not given. To improve the error handling in
this case we check if the repository of the head is the upstream in the
event of a pull request.